Make IResilienceStrategy an abstract class
