+++ src/Polly.Core.Tests/ResilienceStrategyTests.Async.ValueTaskT.cs
@@ -4,41 +4,43 @@ using Xunit;
 
 namespace Polly.Core.Tests;
 
-public partial class ResilienceStrategyExtensionsTests
+#pragma warning disable CS1998 // Async method lacks 'await' operators and will run synchronously
+
+public partial class ResilienceStrategyTests
 {
-    public static IEnumerable<object[]> ExecuteT_EnsureCorrectBehavior_Data()
+    public static IEnumerable<object[]> ExecuteAsyncT_EnsureCorrectBehavior_Data()
     {
-        return ConvertExecuteParameters(ExecuteT_EnsureCorrectBehavior_ExecuteParameters);
+        return ConvertExecuteParameters(ExecuteAsyncT_EnsureCorrectBehavior_ExecuteParameters);
     }
 
-    private static IEnumerable<ExecuteParameters> ExecuteT_EnsureCorrectBehavior_ExecuteParameters()
+    private static IEnumerable<ExecuteParameters> ExecuteAsyncT_EnsureCorrectBehavior_ExecuteParameters()
     {
         long result = 12345;
 
-        yield return new ExecuteParameters<long>(r => r.Execute(t => result), result)
+        yield return new ExecuteParameters<long>(r => r.ExecuteAsync(async t => result), result)
         {
-            Caption = "ExecuteT_NoCancellation",
+            Caption = "ExecuteAsyncT_NoCancellation",
             AssertContext = AssertResilienceContext,
             AssertContextAfter = AssertContextNotInitialized,
         };
 
-        yield return new ExecuteParameters<long>(r => r.Execute(t => { t.Should().Be(CancellationToken); return result; }, CancellationToken), result)
+        yield return new ExecuteParameters<long>(r => r.ExecuteAsync(async t => { t.Should().Be(CancellationToken); return result; }, CancellationToken), result)
         {
-            Caption = "ExecuteT_Cancellation",
+            Caption = "ExecuteAsyncT_Cancellation",
             AssertContext = AssertResilienceContextAndToken,
             AssertContextAfter = AssertContextNotInitialized,
         };
 
-        yield return new ExecuteParameters<long>(r => r.Execute((_, s) => { s.Should().Be("dummy-state"); return result; }, ResilienceContext.Get(), "dummy-state"), result)
+        yield return new ExecuteParameters<long>(r => r.ExecuteAsync(async (_, s) => { s.Should().Be("dummy-state"); return result; }, ResilienceContext.Get(), "dummy-state"), result)
         {
-            Caption = "ExecuteT_ResilienceContextAndState",
+            Caption = "ExecuteAsyncT_ResilienceContextAndState",
             AssertContext = AssertResilienceContext,
             AssertContextAfter = AssertContextInitialized,
         };
 
         static void AssertResilienceContext(ResilienceContext context)
         {
-            context.IsSynchronous.Should().BeTrue();
+            context.IsSynchronous.Should().BeFalse();
             context.IsVoid.Should().BeFalse();
             context.ResultType.Should().Be(typeof(long));
             context.ContinueOnCapturedContext.Should().BeFalse();
@@ -55,9 +57,9 @@ public partial class ResilienceStrategyExtensionsTests
         static void AssertContextInitialized(ResilienceContext context) => context.IsInitialized.Should().BeTrue();
     }
 
-    [MemberData(nameof(ExecuteT_EnsureCorrectBehavior_Data))]
+    [MemberData(nameof(ExecuteAsyncT_EnsureCorrectBehavior_Data))]
     [Theory]
-    public async Task ExecuteT_Ok(ExecuteParameters parameters)
+    public async Task ExecuteAsyncT_Ok(ExecuteParameters parameters)
     {
         ResilienceContext? context = null;
