+++ src/Polly.Core.Tests/ResilienceStrategyTests.Sync.cs
@@ -1,48 +1,44 @@
+using System.Threading.Tasks;
 using FluentAssertions;
 using Polly.Core.Tests.Utils;
 using Xunit;
 
 namespace Polly.Core.Tests;
 
-#pragma warning disable CS1998 // Async method lacks 'await' operators and will run synchronously
-
-public partial class ResilienceStrategyExtensionsTests
+public partial class ResilienceStrategyTests
 {
-    public static IEnumerable<object[]> ExecuteAsTaskAsyncT_EnsureCorrectBehavior_Data()
+    public static IEnumerable<object[]> Execute_EnsureCorrectBehavior_Data()
     {
-        return ConvertExecuteParameters(ExecuteAsTaskAsyncT_EnsureCorrectBehavior_ExecuteParameters);
+        return ConvertExecuteParameters(Execute_EnsureCorrectBehavior_ExecuteParameters);
     }
 
-    private static IEnumerable<ExecuteParameters> ExecuteAsTaskAsyncT_EnsureCorrectBehavior_ExecuteParameters()
+    private static IEnumerable<ExecuteParameters> Execute_EnsureCorrectBehavior_ExecuteParameters()
     {
-        long result = 12345;
-
-        yield return new ExecuteParameters<long>(r => r.ExecuteAsTaskAsync(async t => result), result)
+        yield return new ExecuteParameters(r => r.Execute(_ => { }))
         {
-            Caption = "ExecuteAsTaskAsyncT_NoCancellation",
+            Caption = "Execute_NoCancellation",
             AssertContext = AssertResilienceContext,
             AssertContextAfter = AssertContextNotInitialized,
         };
 
-        yield return new ExecuteParameters<long>(r => r.ExecuteAsTaskAsync(async t => { t.Should().Be(CancellationToken); return result; }, CancellationToken), result)
+        yield return new ExecuteParameters(r => r.Execute(t => { t.Should().Be(CancellationToken); }, CancellationToken))
         {
-            Caption = "ExecuteAsTaskAsyncT_Cancellation",
+            Caption = "Execute_Cancellation",
             AssertContext = AssertResilienceContextAndToken,
             AssertContextAfter = AssertContextNotInitialized,
         };
 
-        yield return new ExecuteParameters<long>(r => r.ExecuteAsTaskAsync(async (_, s) => { s.Should().Be("dummy-state"); return result; }, ResilienceContext.Get(), "dummy-state"), result)
+        yield return new ExecuteParameters(r => r.Execute((_, s) => { s.Should().Be("dummy-state"); }, ResilienceContext.Get(), "dummy-state"))
         {
-            Caption = "ExecuteAsTaskAsyncT_ResilienceContextAndState",
+            Caption = "Execute_ResilienceContextAndState",
             AssertContext = AssertResilienceContext,
             AssertContextAfter = AssertContextInitialized,
         };
 
         static void AssertResilienceContext(ResilienceContext context)
         {
-            context.IsSynchronous.Should().BeFalse();
-            context.IsVoid.Should().BeFalse();
-            context.ResultType.Should().Be(typeof(long));
+            context.IsSynchronous.Should().BeTrue();
+            context.IsVoid.Should().BeTrue();
             context.ContinueOnCapturedContext.Should().BeFalse();
         }
 
@@ -57,9 +53,9 @@ public partial class ResilienceStrategyExtensionsTests
         static void AssertContextInitialized(ResilienceContext context) => context.IsInitialized.Should().BeTrue();
     }
 
-    [MemberData(nameof(ExecuteAsTaskAsyncT_EnsureCorrectBehavior_Data))]
+    [MemberData(nameof(Execute_EnsureCorrectBehavior_Data))]
     [Theory]
-    public async Task ExecuteAsTaskAsyncT_Ok(ExecuteParameters parameters)
+    public async Task Execute_Ok(ExecuteParameters parameters)
     {
         ResilienceContext? context = null;
