+++ src/Polly.Core/NullResilienceStrategy.cs
@@ -3,7 +3,7 @@ namespace Polly;
 /// <summary>
 /// A resilience strategy that executes an user-provided callback without any additional logic.
 /// </summary>
-public sealed class NullResilienceStrategy : IResilienceStrategy
+public sealed class NullResilienceStrategy : ResilienceStrategy
 {
     /// <summary>
     /// Gets the singleton instance of the <see cref="NullResilienceStrategy"/>.
@@ -15,7 +15,7 @@ public sealed class NullResilienceStrategy : IResilienceStrategy
     }
 
     /// <inheritdoc/>
-    ValueTask<TResult> IResilienceStrategy.ExecuteInternalAsync<TResult, TState>(Func<ResilienceContext, TState, ValueTask<TResult>> callback, ResilienceContext context, TState state)
+    protected internal override ValueTask<TResult> ExecuteCoreAsync<TResult, TState>(Func<ResilienceContext, TState, ValueTask<TResult>> callback, ResilienceContext context, TState state)
     {
         Guard.NotNull(callback);
         Guard.NotNull(context);
