+++ src/Polly.Core/ResilienceStrategy.Async.TaskT.cs
@@ -1,30 +1,27 @@
 namespace Polly;
 
-public static partial class ResilienceStrategyExtensions
+public abstract partial class ResilienceStrategy
 {
     /// <summary>
     /// Executes the specified callback.
     /// </summary>
     /// <typeparam name="TResult">The type of result returned by the callback.</typeparam>
     /// <typeparam name="TState">The type of state associated with the callback.</typeparam>
-    /// <param name="strategy">The instance of <see cref="IResilienceStrategy"/>.</param>
     /// <param name="callback">The user-provided callback.</param>
     /// <param name="context">The context associated with the callback.</param>
     /// <param name="state">The state associated with the callback.</param>
     /// <returns>The instance of <see cref="Task"/> that represents the asynchronous execution.</returns>
-    public static async Task<TResult> ExecuteAsTaskAsync<TResult, TState>(
-        this IResilienceStrategy strategy,
+    public async Task<TResult> ExecuteAsTaskAsync<TResult, TState>(
         Func<ResilienceContext, TState, Task<TResult>> callback,
         ResilienceContext context,
         TState state)
     {
-        Guard.NotNull(strategy);
         Guard.NotNull(callback);
         Guard.NotNull(context);
 
         InitializeAsyncContext<TResult>(context);
 
-        return await strategy.ExecuteInternalAsync(
+        return await ExecuteCoreAsync(
             static async (context, state) =>
             {
                 return await state.callback(context, state.state).ConfigureAwait(context.ContinueOnCapturedContext);
@@ -37,23 +34,20 @@ public static partial class ResilienceStrategyExtensions
     /// Executes the specified callback.
     /// </summary>
     /// <typeparam name="TResult">The type of result returned by the callback.</typeparam>
-    /// <param name="strategy">The instance of <see cref="IResilienceStrategy"/>.</param>
     /// <param name="callback">The user-provided callback.</param>
     /// <param name="cancellationToken">The <see cref="CancellationToken"/> associated with the callback.</param>
     /// <returns>The instance of <see cref="ValueTask"/> that represents the asynchronous execution.</returns>
-    public static async Task<TResult> ExecuteAsTaskAsync<TResult>(
-        this IResilienceStrategy strategy,
+    public async Task<TResult> ExecuteAsTaskAsync<TResult>(
         Func<CancellationToken, Task<TResult>> callback,
         CancellationToken cancellationToken = default)
     {
-        Guard.NotNull(strategy);
         Guard.NotNull(callback);
 
         var context = GetAsyncContext<TResult>(cancellationToken);
 
         try
         {
-            return await strategy.ExecuteInternalAsync(
+            return await ExecuteCoreAsync(
                 static async (context, state) =>
                 {
                     return await state(context.CancellationToken).ConfigureAwait(context.ContinueOnCapturedContext);
